truly split to 2 programms
пупупу
