+++ Lab1C/lab1.h
@@ -1,0 +1,9 @@
+#ifndef LAB1_H
+#define LAB1_H
+
+#include <stdio.h>
+#include "array.h"
+
+Array* array_create_and_read(FILE *input);
+
+#endif
