+++ Lab1C/task1.c
@@ -1,5 +1,6 @@
 #include <stdio.h>
 #include "array.h" 
+#include "lab1.h"
 
 void task1(Array *arr)
 {
@@ -12,4 +13,15 @@ void task1(Array *arr)
     }
     //printf("positive: %d\nzeros: %d\nnegative: %d\n", p, z, n);
     printf("+: %d\n0: %d\n-: %d\n", p, z, n);
+}
+
+int main(int argc, char **argv)
+{
+    Array *arr = NULL;
+    FILE *input = fopen(argv[1], "r");
+    arr = array_create_and_read(input);
+    task1(arr);
+    array_delete(arr);
+    fclose(input);
+    return 0;
 }
