+++ Lab1C/tests/test6.txt
@@ -1,4 +1,4 @@
-3
-0 0 0
+6
+-1 -43 -4 -2 -23 -5
 6
 1 2 1 3 4 4
