+++ Lab1C/tests/test7.txt
@@ -1,4 +1,4 @@
-3
-0 0 0
+6
+1 1 2 3 4 4
 6
 1001 2 1 3 4 4 1001
