+++ Lab1C/tests/test8.txt
@@ -1,4 +1,4 @@
 3
-0 0 0
+3 1 32
 6
 -21 2 1 3 4 4 100
